Adds a loop property to SerialTransport.
Retrieve the event loop from the transport rather than relying on the vagaries of get_event_loop().
Tidies up the tests and example code to make it a bit clearer when the transport is set on the
protocol and to avoid linting issues.  Demonstrate how to retrieve the event loop from the
transport in the documentation.